Move out block_registry to break circular dependency.
