+++ io_scene_nif/modules/nif_export/animation/material.py
@@ -40,7 +40,7 @@
 from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules.nif_export.animation import Animation
-from io_scene_nif.modules.nif_export.object.block_registry import block_store
+from io_scene_nif.modules.nif_export.block_registry import block_store
 from io_scene_nif.utils import util_math
 from io_scene_nif.utils.util_global import NifOp
 from io_scene_nif.utils.util_logging import NifLog
