+++ io_scene_nif/modules/nif_export/animation/morph.py
@@ -43,7 +43,7 @@ from pyffi.formats.egm import EgmFormat
 from io_scene_nif.modules.nif_export.animation import Animation
 from io_scene_nif.utils.util_global import EGMData
 
-from io_scene_nif.modules.nif_export.object.block_registry import block_store
+from io_scene_nif.modules.nif_export.block_registry import block_store
 from io_scene_nif.utils.util_global import NifOp
 from io_scene_nif.utils.util_logging import NifLog
