+++ io_scene_nif/modules/nif_export/block_registry.py
