+++ io_scene_nif/modules/nif_export/object/types.py
@@ -37,7 +37,7 @@
 #
 # ***** END LICENSE BLOCK *****
 
-from io_scene_nif.modules.nif_export.object.block_registry import block_store
+from io_scene_nif.modules.nif_export.block_registry import block_store
 from io_scene_nif.utils import util_math
 from io_scene_nif.utils.util_global import NifOp
