+++ io_scene_nif/modules/nif_export/property/texture/writer.py
@@ -42,7 +42,7 @@ import os.path
 import bpy
 from pyffi.formats.nif import NifFormat
 
-from io_scene_nif.modules.nif_export.object.block_registry import block_store
+from io_scene_nif.modules.nif_export.block_registry import block_store
 from io_scene_nif.utils import util_math
 from io_scene_nif.utils.util_global import NifOp
 from io_scene_nif.utils.util_logging import NifLog
