+++ io_scene_nif/nif_export.py
@@ -47,7 +47,7 @@ from pyffi.formats.nif import NifFormat
 from io_scene_nif.modules.nif_export.animation.transform import TransformAnimation
 from io_scene_nif.modules.nif_export.collision import Collision
 from io_scene_nif.modules.nif_export.constraint import Constraint
-from io_scene_nif.modules.nif_export.object.block_registry import block_store
+from io_scene_nif.modules.nif_export.block_registry import block_store
 from io_scene_nif.modules.nif_export.object import Object
 from io_scene_nif.modules.nif_export import scene
 from io_scene_nif.modules.nif_export.property.object import ObjectProperty
